refactor: move app service into domain service in module article
